feature: added footer
